+++ services/pic-sure-hpds/etl/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/etl/phenotype/CSVLoader.java
@@ -66,7 +66,7 @@ public class CSVLoader {
 			String numericValue = record.get(NUMERIC_VALUE);
 			if((numericValue==null || numericValue.isEmpty()) && textValueFromRow!=null) {
 				try {
-					numericValue = Float.parseFloat(textValueFromRow) + "";
+					numericValue = Double.parseDouble(textValueFromRow) + "";
 				}catch(NumberFormatException e) {
 					
 				}
@@ -77,17 +77,17 @@ public class CSVLoader {
 				try {
 					currentConcept[0] = store.store.get(conceptPath);
 				} catch(InvalidCacheLoadException e) {
-					currentConcept[0] = new PhenoCube(conceptPath, isAlpha ? String.class : Float.class);
+					currentConcept[0] = new PhenoCube(conceptPath, isAlpha ? String.class : Double.class);
 					store.store.put(conceptPath, currentConcept[0]);
 				}
 			}
 			String value = isAlpha ? record.get(TEXT_VALUE) : numericValue;
 
-			if(value != null && !value.trim().isEmpty() && ((isAlpha && currentConcept[0].vType == String.class)||(!isAlpha && currentConcept[0].vType == Float.class))) {
+			if(value != null && !value.trim().isEmpty() && ((isAlpha && currentConcept[0].vType == String.class)||(!isAlpha && currentConcept[0].vType == Double.class))) {
 				value = value.trim();
-				currentConcept[0].setColumnWidth(isAlpha ? Math.max(currentConcept[0].getColumnWidth(), value.getBytes().length) : Float.BYTES);
+				currentConcept[0].setColumnWidth(isAlpha ? Math.max(currentConcept[0].getColumnWidth(), value.getBytes().length) : Double.BYTES);
 				int patientId = Integer.parseInt(record.get(PATIENT_NUM));
-				currentConcept[0].add(patientId, isAlpha ? value : Float.parseFloat(value));
+				currentConcept[0].add(patientId, isAlpha ? value : Double.parseDouble(value));
 				store.allIds.add(patientId);
 			}
 		} catch (ExecutionException e) {
